[skip ci] remove SimpleCViewAsync
